Adding add_node_end and print_list functions into --useful_func.c
